um: remove mmap_sem usage from activate_mm()
Lockdep complains about the mmap_sem because we try to
down_write() it while we already hold the task (spin)lock.

Since commit 9bfb23fc4a48 ("sys_unshare: remove the dead
CLONE_THREAD/SIGHAND/VM code") this is no longer called
by sys_unshare(), and thus the only caller of this is
exec_mmap() in fs/exec.c. But in this case, the new mm
is completely new, so it shouldn't be possible to change
it concurrently.

Remove the mmap_sem handling completely.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>